@@ -25,12 +25,11 @@
   },
   "homepage": "https://github.com/magicdawn/node-mysql-aes#readme",
   "devDependencies": {
-    "co-mocha": "^1.1.3",
-    "codecov": "^2.1.0",
-    "eslint": "^4.1.0",
-    "istanbul": "^0.4.5",
-    "mocha": "^3.0.2",
-    "should": "^12.0.0"
+    "codecov": "^3.8.2",
+    "eslint": "^8.26.0",
+    "mocha": "^10.1.0",
+    "nyc": "^15.1.0",
+    "should": "^13.2.3"
   },
   "files": [
     "index.js",
